+++ src/app/Charts/divergingBarChart.js
