+++ src/app/Page-Functionality/Login/auth-context.js
@@ -2,7 +2,7 @@
 
 import { createContext } from "react"
 
-import {auth} from './index'
+import {auth} from '../../index'
 import { GoogleAuthProvider, FacebookAuthProvider, signInWithPopup, signOut } from "firebase/auth"
 import { useAuthState } from 'react-firebase-hooks/auth';
