+++ src/app/layout.js
@@ -1,9 +1,9 @@
 'use client'
 import "./globals.css";
-import Nav from './Navigations'
-import AuthContextProvider from "./auth-context";
+import Nav from './Page-Functionality/Navigations'
+import AuthContextProvider from "./Page-Functionality/Login/auth-context";
 
-import FinanceContextProvider from "./finance-context";
+import FinanceContextProvider from "./Finance-Context/finance-context";
 import { ToastContainer, toast} from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
